+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/web/controller/MeaningfulUseController.java
@@ -4,17 +4,13 @@ import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.util.Date;
-import java.util.LinkedHashSet;
 import java.util.List;
-import java.util.Set;
 
-import org.apache.commons.csv.CSVFormat;
-import org.apache.commons.csv.CSVParser;
-import org.apache.commons.csv.CSVRecord;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.util.StringUtils;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -24,17 +20,22 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MaxUploadSizeExceededException;
 import org.springframework.web.multipart.MultipartFile;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-
+import gov.healthit.chpl.auth.Util;
+import gov.healthit.chpl.auth.dto.UserDTO;
+import gov.healthit.chpl.auth.manager.UserManager;
+import gov.healthit.chpl.auth.user.UserRetrievalException;
 import gov.healthit.chpl.dao.EntityCreationException;
 import gov.healthit.chpl.dao.EntityRetrievalException;
 import gov.healthit.chpl.domain.AccurateAsOfDate;
-import gov.healthit.chpl.domain.MeaningfulUseUser;
+import gov.healthit.chpl.domain.Job;
+import gov.healthit.chpl.domain.concept.JobTypeConcept;
 import gov.healthit.chpl.dto.MeaningfulUseAccurateAsOfDTO;
+import gov.healthit.chpl.dto.job.JobDTO;
+import gov.healthit.chpl.dto.job.JobTypeDTO;
 import gov.healthit.chpl.manager.CertifiedProductManager;
+import gov.healthit.chpl.manager.JobManager;
 import gov.healthit.chpl.manager.MeaningfulUseManager;
 import gov.healthit.chpl.web.controller.exception.ValidationException;
-import gov.healthit.chpl.web.controller.results.MeaningfulUseUserResults;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 
@@ -43,110 +44,95 @@ import io.swagger.annotations.ApiOperation;
 @RequestMapping("/meaningful_use")
 public class MeaningfulUseController {
 	private static final Logger logger = LogManager.getLogger(MeaningfulUseController.class);
-	
+	private final JobTypeConcept allowedJobType = JobTypeConcept.MUU_UPLOAD;
 	@Autowired CertifiedProductManager cpManager;
 	@Autowired MeaningfulUseManager muManager;
+	@Autowired JobManager jobManager;
+	@Autowired UserManager userManager;
 	
 	@ApiOperation(value="Upload a file to update the number of meaningful use users for each CHPL Product Number", 
 			notes="Accepts a CSV file with chpl_product_number and num_meaningful_use_users to update the number of meaningful use users for each CHPL Product Number."
 					+ " The user uploading the file must have ROLE_ADMIN or ROLE_ONC_STAFF ")
 	@RequestMapping(value="/upload", method=RequestMethod.POST,
 			produces="application/json; charset=utf-8") 
-	public @ResponseBody MeaningfulUseUserResults uploadMeaningfulUseUsers(@RequestParam("file") MultipartFile file) throws ValidationException, MaxUploadSizeExceededException {
+	public synchronized ResponseEntity<Job> uploadMeaningfulUseUsers(@RequestParam("file") MultipartFile file) 
+			throws EntityRetrievalException, EntityCreationException, 
+			ValidationException, MaxUploadSizeExceededException {
+		
+		if(Util.getCurrentUser() == null || Util.getCurrentUser().getId() == null) {
+			return new ResponseEntity<Job>(HttpStatus.UNAUTHORIZED);
+		}
+		
 		if (file.isEmpty()) {
 			throw new ValidationException("You cannot upload an empty file!");
 		}
 		
+		//for now we'll only accept CSV
 		if(!file.getContentType().equalsIgnoreCase("text/csv") &&
 				!file.getContentType().equalsIgnoreCase("application/vnd.ms-excel")) {
 			throw new ValidationException("File must be a CSV document.");
 		}
-		MeaningfulUseUserResults meaningfulUseUserResults = new MeaningfulUseUserResults();
-		Set<MeaningfulUseUser> muusToUpdate = new LinkedHashSet<MeaningfulUseUser>();
-		Set<String> uniqueMuusFromFile = new LinkedHashSet<String>();
 		
-		BufferedReader reader = null;
-		CSVParser parser = null;
+		//figure out the user
+		UserDTO currentUser = null;
 		try {
-			reader = new BufferedReader(new InputStreamReader(file.getInputStream()));
-			parser = new CSVParser(reader, CSVFormat.EXCEL);
-			
-			List<CSVRecord> records = parser.getRecords();
-			if(records.size() <= 1) {
-				throw new ValidationException("The file appears to have a header line with no other information. Please make sure there are at least two rows in the CSV file.");
-			}
-			
-			CSVRecord heading = null;
-			
-			for(int i = 1; i <= records.size(); i++){
-				CSVRecord currRecord = records.get(i-1);
-				MeaningfulUseUser muu = new MeaningfulUseUser();
-				
-				// add header if something similar to "chpl_product_number" and "num_meaningful_use" exists
-				if(heading == null && i == 1 && !StringUtils.isEmpty(currRecord.get(0).trim()) && currRecord.get(0).trim().contains("product")
-						&& !StringUtils.isEmpty(currRecord.get(1).trim()) && currRecord.get(1).trim().contains("meaning")) {
-					heading = currRecord;
-				}
-				// populate MeaningfulUseUserResults
-				else {
-					String chplProductNumber = currRecord.get(0).trim();
-					Long numMeaningfulUseUsers = null;
-					try{
-						numMeaningfulUseUsers = Long.parseLong(currRecord.get(1).trim());
-						muu.setProductNumber(chplProductNumber);
-						muu.setNumberOfUsers(numMeaningfulUseUsers);
-						muu.setCsvLineNumber(i);
-						// check if product number has already been updated
-						if(uniqueMuusFromFile.contains(muu.getProductNumber())){
-							throw new IOException();
-						}
-						muusToUpdate.add(muu);
-						uniqueMuusFromFile.add(muu.getProductNumber());
-					} catch (NumberFormatException e){
-						muu.setProductNumber(chplProductNumber);
-						muu.setCsvLineNumber(i);
-						muu.setError("Line " + muu.getCsvLineNumber() + ": Field \"num_meaningful_use\" with value \"" + currRecord.get(1).trim() + "\" is invalid. "
-								+ "Value in field \"num_meaningful_use\" must be an integer.");
-						muusToUpdate.add(muu);
-						uniqueMuusFromFile.add(muu.getProductNumber());
-					}
-					catch (IOException e){
-						muu.setProductNumber(chplProductNumber);
-						muu.setCsvLineNumber(i);
-						Integer dupLineNumber = null;
-						// get line number with duplicate chpl_product_number
-						for (MeaningfulUseUser entry: muusToUpdate) {
-						     if (entry.getProductNumber().equals(muu.getProductNumber())){
-						    	 dupLineNumber = entry.getCsvLineNumber();
-						     }
-						   }
-						muu.setError("Line " + muu.getCsvLineNumber() + ": Field \"chpl_product_number\" with value \"" + muu.getProductNumber() + "\" is invalid. "
-								+ "Duplicate \"chpl_product_number\" at line " + dupLineNumber);
-						muusToUpdate.add(muu);
-					}
-				}
-			}
-		} catch(IOException ioEx) {
-			logger.error("Could not get input stream for uploaded file " + file.getName());			
-			throw new ValidationException("Could not get input stream for uploaded file " + file.getName());
-		} finally {
-			 try { parser.close(); } catch(Exception ignore) {}
-			try { reader.close(); } catch(Exception ignore) {}
+			currentUser = userManager.getById(Util.getCurrentUser().getId());
+		} catch(UserRetrievalException ex) {
+			logger.error("Error finding user with ID " + Util.getCurrentUser().getId() + ": " + ex.getMessage());
+			return new ResponseEntity<Job>(HttpStatus.UNAUTHORIZED);
+		}
+		if(currentUser == null) {
+			logger.error("No user with ID " + Util.getCurrentUser().getId() + " could be found in the system.");
+			return new ResponseEntity<Job>(HttpStatus.UNAUTHORIZED);
 		}
 		
-		try {
-			meaningfulUseUserResults = cpManager.updateMeaningfulUseUsers(muusToUpdate);
-		} catch (JsonProcessingException e) {
-			e.printStackTrace();
-		} catch (EntityCreationException e) {
-			e.printStackTrace();
-		} catch (EntityRetrievalException e) {
-			e.printStackTrace();
-		} catch (IOException e) {
-			e.printStackTrace();
-		} 
+		JobTypeDTO jobType = null;
+		List<JobTypeDTO> jobTypes = jobManager.getAllJobTypes();
+		for(JobTypeDTO jt : jobTypes) {
+			if(jt.getName().equalsIgnoreCase(allowedJobType.getName())) {
+				jobType = jt;
+			}
+		}
 		
-		return meaningfulUseUserResults;
+		//read the file into a string
+		StringBuffer data = new StringBuffer();
+		BufferedReader reader = null;
+		try {
+			reader = new BufferedReader(new InputStreamReader(file.getInputStream()));
+			String line = null;
+			while((line = reader.readLine()) != null) {
+				if(data.length() > 0) {
+					data.append(System.getProperty("line.separator"));
+				}
+				data.append(line);
+			}		
+		} catch(IOException ex) {
+			String msg = "Could not read file: " + ex.getMessage();
+			logger.error(msg);
+			throw new ValidationException(msg);
+		}
+		
+		JobDTO toCreate = new JobDTO();
+		toCreate.setData(data.toString());
+		toCreate.setUser(currentUser);
+		toCreate.setJobType(jobType);
+		JobDTO insertedJob = jobManager.createJob(toCreate);
+		JobDTO createdJob = jobManager.getJobById(insertedJob.getId());
+		
+		try {
+			boolean isStarted = jobManager.start(createdJob);
+			if(!isStarted) {
+				return new ResponseEntity<Job>(new Job(createdJob), HttpStatus.BAD_REQUEST);
+			} else {
+				createdJob = jobManager.getJobById(insertedJob.getId());
+			}
+		} catch(EntityRetrievalException ex) {
+			logger.error("Could not mark job " + createdJob.getId() + " as started.");
+			return new ResponseEntity<Job>(new Job(createdJob), HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+		
+		//query the now running job
+		return new ResponseEntity<Job>(new Job(createdJob), HttpStatus.OK);
 	}
 	
 	@ApiOperation(value="Get a single date value to indicate when the last meaningful use user file data is accurate as of.",